Move FakeIOHost
