+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/private/testing/index.ts
@@ -1,1 +1,2 @@
 export * from './test-io-host';
+export * from './fake-io-host';
